// legacy/frontend/app/(main)/pages/practice/types.ts
export interface IUser {
  id?: number;
  name?: string;
  email?: string;
  email_verified?: string | null; // Assuming datetime is converted to string
  image?: string;
  hash?: string;
  sr_alg_type?: string;
  phone?: string;
  phone_verified?: string | null; // Assuming datetime is converted to string
  acceptable_delinquency_window?: number;
}